update: complete decomposition
